requirements.txt
streamlit==1.35.0
numpy==2.4.3
pandas==3.0.1
scikit-learn==1.5.0
google-genai==0.3.0